fix(audiobridge): str_dup the UUIDs for users and rooms
Previously, the room_uuid_string and user_uuid_string did not duplicate
the strings. This would result in intermittent failures, depending on
the value of the pointer at the time of it being used.

If the string value of the pointer contained non-ascii characters, the
SDP answer would fail with no clear indication of the reason why.