feat(audiobridge): allow audio interop over transport
The audio stream of a room participant is now broadcast over the
transport using base64 encoding. The base64 encoding keeps the data
transport agnostic.

It is also possible to add `external` participants, allowing base64 OPUS
frames to be sent to the channel to be mixed with the participants in
the room.